Added a shim for a default registry to be shared by the asset importer and the stencil for the GTF editor-- this ensures a context node exists in CLDS whenever a graph is created-- this will need to be explored further next quarter when targets and context nodes are explored proper.

+++ com.unity.shadergraph/Editor/GraphDeltaRegistry/Generation/Interpreter.cs
@@ -129,11 +129,11 @@ namespace UnityEditor.ShaderGraph.Generation
         {
             //length by height
             string lxh = "";
-            if(entry.length > 1 || entry.height > 1)
+            if((int)entry.length > 1 || (int)entry.height > 1)
             {
                 lxh += entry.length;
             }
-            if(entry.height > 1)
+            if((int)entry.height > 1)
             {
                 lxh += "x" + entry.height;
             }
